pyproject.toml
[tool.poetry]
name = "job_gear_sample"
version = "0.1.0"
description = "Sample of a implementation of a activite executable in a background pipline"
authors = ["Marlos Ribeiro"]

[tool.poetry.dependencies]
python = "3.9"
celery = "^5.2.7"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7"
pylint = "^2.14.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
